Provide simple PSR-16 cache implementation from configuration
Summary:
Cache with basic get/set/has/delete/clear operations.
Abstraction layer for Memcached, Redis and a void mock-purpose.

Similar to the database package, read a configuration array
to build and load a concrete cache implementation.

This library is a PSR-16 implementation.
This library does NOT currently support PSR-6.

This library is intended to replace the Zed Cache engine,
and is based on the includes/cache classes from Zed code.

Test Plan: Full units test coverage

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3233